% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buat_skripsi.R
\name{render_skripsi}
\alias{render_skripsi}
\title{Render proyek skripsi menggunakan Quarto}
\usage{
render_skripsi(path = ".")
}
\arguments{
\item{path}{Direktori proyek skripsi.}
}
\value{
(Secara \emph{invisible}) \code{path}.
}
\description{
Pembungkus tipis untuk menjalankan \verb{quarto render} pada direktori proyek.
}
