\docType{package}
\name{skripsistis-package}
\alias{skripsistis}
\alias{skripsistis-package}
\title{skripsistis: Template Quarto Skripsi Politeknik Statistika STIS}
\description{
Kerangka proyek Quarto untuk penyusunan skripsi di Politeknik Statistika STIS
sesuai Pedoman Skripsi KS Edisi Keenam (2025), dengan gaya kutipan APA edisi
ke-7 modifikasi bahasa Indonesia.
}
\seealso{
\code{\link{buat_skripsi}}, \code{\link{render_skripsi}}, \code{\link{path_csl}}
}
